Add the Logger component
This particular commit adds an implementation of the system-wide logger component, which
is used by the developers to write warnings, error information or just messages into
the specified file. The component is not tread-safe and is meant to be a first version
of the component (will be updated in the future).