Implement a lexer
This commit implements a lexer component, which is used to split text
into a set of labeled tokens, which will be then processed by the
parser.

Along with the lexer itself, essential exceptions (like
UnexpectedLexemeException), handlers and tokens were implemented.

This commit also brings changes to the hierarchy of the unit tests,
covering the new components and fixing the tests for the old ones,
changed in order to assure proper lexer implementation.