Implement a parser
This commit implements a parser component, which is used to properly
handle every possible token, got from the lexer, and react on
unexpected ones,

Along with the parser itself, essential exceptions were implemented.

This commit also brings changes to the hierarchy of the unit tests,
covering the new components and fixing the tests for the old ones,
changed in order to assure proper parser implementation.