Separate keyword state from statements
This commit creates a separate component, used as a parser state named
"Keyword". This one incapsulates all the keyword state-selection routine
cleaning the "Statement" state and leaving it with only EOF concerns.

There is also a new exception that is used to indicate where an unknown
keyword is encountered and how it looks like.